moved EntityExtensions to EntityModels folder. updated all references. delete old Entities folder
