--- Fabric.Identity.API/Persistence/SqlServer/Entities/IdentityResourceEntity.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Fabric.Identity.API.Persistence.SqlServer.EntityModels;
-
-namespace Fabric.Identity.API.Persistence.SqlServer.Entities
-{
-    public class IdentityResourceEntity : IdentityServer4.EntityFramework.Entities.IdentityResource, ITrackable, ISoftDelete
-    {
-        public DateTime CreatedDateTimeUtc { get; set; }
-        public DateTime? ModifiedDateTimeUtc { get; set; }
-        public string CreatedBy { get; set; }
-        public string ModifiedBy { get; set; }
-        public bool IsDeleted { get; set; }
-    }
-}
