--- Fabric.Identity.API/Persistence/SqlServer/Entities/PersistedGrantEntity.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Fabric.Identity.API.Persistence.SqlServer.EntityModels;
-
-namespace Fabric.Identity.API.Persistence.SqlServer.Entities
-{
-    public class PersistedGrantEntity : IdentityServer4.EntityFramework.Entities.PersistedGrant, ISoftDelete, ITrackable
-    {
-        public bool IsDeleted { get; set; }
-        public DateTime CreatedDateTimeUtc { get; set; }
-        public DateTime? ModifiedDateTimeUtc { get; set; }
-        public string CreatedBy { get; set; }
-        public string ModifiedBy { get; set; }
-    }
-}
