+++ Fabric.Identity.API/Persistence/SqlServer/ModelBuilderExtensions.cs
@@ -1,13 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Fabric.Identity.API.Persistence.SqlServer.Entities;
-using IdentityServer4.EntityFramework.Entities;
-using IdentityServer4.EntityFramework.Mappers;
-using IdentityServer4.EntityFramework.Options;
+﻿using IdentityServer4.EntityFramework.Options;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Fabric.Identity.API.Persistence.SqlServer
@@ -20,89 +12,7 @@ namespace Fabric.Identity.API.Persistence.SqlServer
             return string.IsNullOrWhiteSpace(configuration.Schema) ? entityTypeBuilder.ToTable(configuration.Name) : entityTypeBuilder.ToTable(configuration.Name, configuration.Schema);
         }
 
-        /// <summary>
-        /// Configures the client context.
-        /// </summary>
-        /// <param name="modelBuilder">The model builder.</param>
-        /// <param name="storeOptions">The store options.</param>
-        public static void ConfigureClientContext(this ModelBuilder modelBuilder, ConfigurationStoreOptions storeOptions)
-        {
-            if (!string.IsNullOrWhiteSpace(storeOptions.DefaultSchema)) modelBuilder.HasDefaultSchema(storeOptions.DefaultSchema);
-
-            modelBuilder.Entity<ClientEntity>(client =>
-            {
-                client.ToTable(storeOptions.Client);
-                client.HasKey(x => x.Id);
-
-                client.Property(x => x.ClientId).HasMaxLength(200).IsRequired();
-                client.Property(x => x.ProtocolType).HasMaxLength(200).IsRequired();
-                client.Property(x => x.ClientName).HasMaxLength(200);
-                client.Property(x => x.ClientUri).HasMaxLength(2000);
-                client.Property(x => x.LogoUri).HasMaxLength(2000);
-
-                client.HasIndex(x => x.ClientId).IsUnique();
-
-                client.HasMany(x => x.AllowedGrantTypes);
-                client.HasMany(x => x.RedirectUris);
-                client.HasMany(x => x.PostLogoutRedirectUris);
-                client.HasMany(x => x.AllowedScopes);
-                client.HasMany(x => x.ClientSecrets);
-                client.HasMany(x => x.Claims);
-                client.HasMany(x => x.IdentityProviderRestrictions);
-                client.HasMany(x => x.AllowedCorsOrigins);
-            });
-
-            modelBuilder.Entity<ClientGrantType>(grantType =>
-            {
-                grantType.ToTable(storeOptions.ClientGrantType);
-                grantType.Property(x => x.GrantType).HasMaxLength(250).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientRedirectUri>(redirectUri =>
-            {
-                redirectUri.ToTable(storeOptions.ClientRedirectUri);
-                redirectUri.Property(x => x.RedirectUri).HasMaxLength(2000).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientPostLogoutRedirectUri>(postLogoutRedirectUri =>
-            {
-                postLogoutRedirectUri.ToTable(storeOptions.ClientPostLogoutRedirectUri);
-                postLogoutRedirectUri.Property(x => x.PostLogoutRedirectUri).HasMaxLength(2000).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientScope>(scope =>
-            {
-                scope.ToTable(storeOptions.ClientScopes);
-                scope.Property(x => x.Scope).HasMaxLength(200).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientSecret>(secret =>
-            {
-                secret.ToTable(storeOptions.ClientSecret);
-                secret.Property(x => x.Value).HasMaxLength(2000).IsRequired();
-                secret.Property(x => x.Type).HasMaxLength(250);
-                secret.Property(x => x.Description).HasMaxLength(2000);
-            });
-
-            modelBuilder.Entity<ClientClaim>(claim =>
-            {
-                claim.ToTable(storeOptions.ClientClaim);
-                claim.Property(x => x.Type).HasMaxLength(250).IsRequired();
-                claim.Property(x => x.Value).HasMaxLength(250).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientIdPRestriction>(idPRestriction =>
-            {
-                idPRestriction.ToTable(storeOptions.ClientIdPRestriction);
-                idPRestriction.Property(x => x.Provider).HasMaxLength(200).IsRequired();
-            });
-
-            modelBuilder.Entity<ClientCorsOrigin>(corsOrigin =>
-            {
-                corsOrigin.ToTable(storeOptions.ClientCorsOrigin);
-                corsOrigin.Property(x => x.Origin).HasMaxLength(150).IsRequired();
-            });          
-        }
+       
 
     }
 }
