+++ Fabric.Identity.API/Persistence/SqlServer/Stores/SqlServerIdentityResourceStore.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Fabric.Identity.API.Persistence.SqlServer.Entities;
 using Fabric.Identity.API.Persistence.SqlServer.Services;
-using IdentityServer4.Models;
 using Microsoft.EntityFrameworkCore;
+using Fabric.Identity.API.Persistence.SqlServer.EntityModels;
+using IdentityResource = IdentityServer4.Models.IdentityResource;
 
 namespace Fabric.Identity.API.Persistence.SqlServer.Stores
 {
